backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# AI/ML Dependencies
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
mediapipe==0.10.9
scipy==1.11.4

# BLE
bleak==0.21.1

# Notifications
sendgrid==6.11.0

# Security
PyJWT==2.8.0
bcrypt==4.1.1

# Utilities
requests==2.31.0
